--- app/Repositories/ReplyRepository.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace App\Repositories;
-
-
-/**
- * Interface ReplyRepository
- * @package namespace App\Repositories;
- */
-interface ReplyRepository extends BaseRepository
-{
-    //
-}
