+++ app/Repositories/ReplyRepositoryInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Repositories;
+
+
+/**
+ * Interface ReplyRepository.
+ */
+interface ReplyRepositoryInterface extends RepositoryInterface
+{
+    //TODO
+}
