--- app/Repositories/TopicRepository.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace App\Repositories;
-
-/**
- * Interface TopicRepository
- * @package namespace App\Repositories;
- */
-interface TopicRepository extends BaseRepository
-{
-
-}
