+++ app/Repositories/TopicRepositoryInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Repositories;
+
+/**
+ * Interface TopicRepositoryInterface.
+ */
+interface TopicRepositoryInterface extends RepositoryInterface
+{
+    //TODO
+}
