--- app/Repositories/UserRepository.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace App\Repositories;
-
-/**
- * Interface UserRepository
- * @package namespace App\Repositories;
- */
-interface UserRepository extends BaseRepository
-{
-    //
-}
