+++ app/Repositories/UserRepositoryInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Repositories;
+
+/**
+ * Interface UserRepositoryInterface.
+ */
+interface UserRepositoryInterface extends RepositoryInterface
+{
+    //
+}
