+++ app/User.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models;
+namespace App;
 
 use App\Presenters\UserPresenter;
 use Illuminate\Database\Eloquent\Model;
